Fold MergePerson CommonPageTests into the page test classes
Journeys don't get a CommonPageTests class; each page's tests live with
that page. The theories parameterised by page are split across
EnterTrnTests, MatchesTests, MergePersonTests and CheckAnswersTests,
each keeping only its own rows.

Three of the enter-trn cases were dropped rather than moved:
EnterTrnTests already had its own Post_PersonAIsDeactivated,
Post_PersonAHasOpenAlert and Post_PersonAHasInvalidInductionStatus tests
covering the same scenarios with the same induction statuses, so the
count goes 190 -> 185 with no loss of coverage.
